contact with errors
